Game of Craps in Java
This simple GUI based program uses the Java Swing layout (Windows look and feel) for the UI and the functionality is quite simple. User clicks a button and rolls the dice pseudorandomly. The outcome of the roll is displayed in a label below. This wasn't meant to be a complicated project. Just something to help with improving my programming skills and learn more about the Java language. Keep in mind this uses a much older version of Java runtime (back in 2010).