Physics Calculator in Java
This calculator is used to solve common physics equations used in high school physics courses. Once a variable of a selected equation is chosen, the user can enter values in for the remaining variables and click calculate in order to get the computational result. Note this project is in a zip format due to exceeding the 100 files limit on Github