Averaging Grades in Java
Basic Java averaging grades calculator using Java swing dialog boxes for input. Enter in the number of grades to start with, After that, enter the score of each grade, then enter the points out of for the assignment. Once these steps are completed, then the calculator will produce a result which is then converted into a final percentage. This is meant to be for individual scores.