Party Room Booker
A small Java program used for booking party rooms. Users enter a name, phone number, and select number of people in the party. They must choose non smoking or smoking to book a room. If a rooms are full for smoking or non smoking, then user is alerted that there are no rooms available for the chosen section. The Java Swing framework was used for implementing the GUI interface.